Include Google Scholar metadata
Added the archive of the Thesis of the group (it will be later migrated
to the research group website).

A page for each thesis and article is now generated. In this page, the
abstract is included, and a certain set of <meta> tags interpreted by
Google Scholar are added.

This change (which increases the generation time by an order of
magnitude!) should eventually fix the whole mess that some conferences
are doing with their shitty templates (one above all: WSC!) in Scholar
bibliographic database.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>